improved figure selection fixed zoom in point selection support up/down scrollbuttons nice icon

+++ tesselation/tline.c
@@ -96,11 +96,11 @@ TLineTesselate(TLinenode_type l1, int x, int y, double scale, double ca,
     TVertexnode_type *otv = l1.rootnode;
     int ox, oy;
     int nx, ny;
-    ox = (otv->p1.x * ca - otv->p1.y * sa) * scale + x;
-    oy = (otv->p1.x * sa + otv->p1.y * ca) * scale + y;
+    ox = (int)((otv->p1.x * ca - otv->p1.y * sa) * scale + x);
+    oy = (int)((otv->p1.x * sa + otv->p1.y * ca) * scale + y);
     while (otv->next != NULL) {
-	nx = (otv->next->p1.x * ca - otv->next->p1.y * sa) * scale + x;
-	ny = (otv->next->p1.x * sa + otv->next->p1.y * ca) * scale + y;
+	nx = (int)((otv->next->p1.x * ca - otv->next->p1.y * sa) * scale + x);
+	ny = (int)((otv->next->p1.x * sa + otv->next->p1.y * ca) * scale + y);
 	WinDrawLine(ox, oy, nx, ny);
 	ox = nx;
 	oy = ny;
